+++ libs/create-langgraph/CHANGELOG.md
@@ -1,5 +1,23 @@
 # create-langgraph
 
+## 1.1.3
+
+### Patch Changes
+
+- [#1853](https://github.com/langchain-ai/langgraphjs/pull/1853) [`a84c1ff`](https://github.com/langchain-ai/langgraphjs/commit/a84c1ff18289653ff4715bd0db4ac3d06600556e) Thanks [@christian-bromann](https://github.com/christian-bromann)! - retry release: allow user to generate a langgraph config file
+
+## 1.1.2
+
+### Patch Changes
+
+- [#1850](https://github.com/langchain-ai/langgraphjs/pull/1850) [`e9f7e8e`](https://github.com/langchain-ai/langgraphjs/commit/e9f7e8e9e6b8851cb7dd68e31d2f1867b62bd6bd) Thanks [@christian-bromann](https://github.com/christian-bromann)! - retry release: allow user to generate a langgraph config file
+
+## 1.1.1
+
+### Patch Changes
+
+- 3ec85a4: retry release: allow user to generate a langgraph config file
+
 ## 1.1.0
 
 ### Minor Changes