+++ libs/langgraph-api/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @langchain/langgraph-api
 
+## 1.1.8
+
+### Patch Changes
+
+- [#1853](https://github.com/langchain-ai/langgraphjs/pull/1853) [`a84c1ff`](https://github.com/langchain-ai/langgraphjs/commit/a84c1ff18289653ff4715bd0db4ac3d06600556e) Thanks [@christian-bromann](https://github.com/christian-bromann)! - retry release: bump @hono/zod-validator
+
+- Updated dependencies []:
+  - @langchain/langgraph-ui@1.1.8
+
+## 1.1.7
+
+### Patch Changes
+
+- [#1850](https://github.com/langchain-ai/langgraphjs/pull/1850) [`e9f7e8e`](https://github.com/langchain-ai/langgraphjs/commit/e9f7e8e9e6b8851cb7dd68e31d2f1867b62bd6bd) Thanks [@christian-bromann](https://github.com/christian-bromann)! - retry release: bump @hono/zod-validator
+
+- Updated dependencies []:
+  - @langchain/langgraph-ui@1.1.7
+
+## 1.1.6
+
+### Patch Changes
+
+- 3ec85a4: retry release: bump @hono/zod-validator
+  - @langchain/langgraph-ui@1.1.6
+
+## 1.1.5
+
+### Patch Changes
+
+- 3613386: retry release: bump @hono/zod-validator
+  - @langchain/langgraph-ui@1.1.5
+
+## 1.1.4
+
+### Patch Changes
+
+- 730dc7c: fix(api): bump @hono/zod-validator
+  - @langchain/langgraph-ui@1.1.4
+
 ## 1.1.3
 
 ### Patch Changes