+++ libs/langgraph-cli/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @langchain/langgraph-cli
 
+## 1.1.8
+
+### Patch Changes
+
+- Updated dependencies [[`a84c1ff`](https://github.com/langchain-ai/langgraphjs/commit/a84c1ff18289653ff4715bd0db4ac3d06600556e), [`a84c1ff`](https://github.com/langchain-ai/langgraphjs/commit/a84c1ff18289653ff4715bd0db4ac3d06600556e)]:
+  - @langchain/langgraph-api@1.1.8
+  - create-langgraph@1.1.3
+
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies [[`e9f7e8e`](https://github.com/langchain-ai/langgraphjs/commit/e9f7e8e9e6b8851cb7dd68e31d2f1867b62bd6bd), [`e9f7e8e`](https://github.com/langchain-ai/langgraphjs/commit/e9f7e8e9e6b8851cb7dd68e31d2f1867b62bd6bd)]:
+  - @langchain/langgraph-api@1.1.7
+  - create-langgraph@1.1.2
+
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [3ec85a4]
+- Updated dependencies [3ec85a4]
+  - @langchain/langgraph-api@1.1.6
+  - create-langgraph@1.1.1
+
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies [3613386]
+  - @langchain/langgraph-api@1.1.5
+
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [730dc7c]
+  - @langchain/langgraph-api@1.1.4
+
 ## 1.1.3
 
 ### Patch Changes