+++ libs/sdk/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @langchain/langgraph-sdk
 
+## 1.4.4
+
+### Patch Changes
+
+- [#1853](https://github.com/langchain-ai/langgraphjs/pull/1853) [`a84c1ff`](https://github.com/langchain-ai/langgraphjs/commit/a84c1ff18289653ff4715bd0db4ac3d06600556e) Thanks [@christian-bromann](https://github.com/christian-bromann)! - retry release: add type-safe tool call streaming with agent type inference, provide proper error message when failing to connect to a server, expose Thread["config"] and Thread["error"]
+
+## 1.4.3
+
+### Patch Changes
+
+- [#1850](https://github.com/langchain-ai/langgraphjs/pull/1850) [`e9f7e8e`](https://github.com/langchain-ai/langgraphjs/commit/e9f7e8e9e6b8851cb7dd68e31d2f1867b62bd6bd) Thanks [@christian-bromann](https://github.com/christian-bromann)! - retry release: add type-safe tool call streaming with agent type inference, provide proper error message when failing to connect to a server, expose Thread["config"] and Thread["error"]
+
+## 1.4.2
+
+### Patch Changes
+
+- 3ec85a4: retry release: add type-safe tool call streaming with agent type inference, provide proper error message when failing to connect to a server, expose Thread["config"] and Thread["error"]
+
+## 1.4.1
+
+### Patch Changes
+
+- 3613386: retry release: add type-safe tool call streaming with agent type inference, provide proper error message when failing to connect to a server, expose Thread["config"] and Thread["error"]
+
+## 1.4.0
+
+### Minor Changes
+
+- 730dc7c: feat(sdk): add type-safe tool call streaming with agent type inference
+
+### Patch Changes
+
+- 730dc7c: fix(sdk): provide proper error message when failing to connect to a server
+- 4ffdde9: Expose `Thread["config"]` and `Thread["error"]`
+
 ## 1.3.1
 
 ### Patch Changes